Move the code out to utils
